// web/src/components/daedongMap/types.ts
export type { BakeryEntity } from '@/remotes/network/fetchBakeries';

export type Bread =
  | '기본'
  | '식사빵'
  | '비건/키토'
  | '크로와상'
  | '구움과자류'
  | '케이크'
  | '파이/타르트'
  | '마카롱'
  | '도넛'
  | '쿠키'
  | '크림빵'
  | '과자류'
  | '추억의빵'
  | '기타';